+++ Warbots2020/src/main/java/frc/robot/dashboard/Update.java
@@ -13,6 +13,11 @@ import frc.robot.drivetrain.*;
 import frc.robot.bling.*;
 import frc.robot.climber.*;
 import frc.robot.shooter.*;
+import frc.robot.subsystems.Climber;
+import frc.robot.subsystems.DriveTrain;
+import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.Loader;
+import frc.robot.subsystems.Shooter;
 import frc.robot.intake.*;
 import frc.robot.loader.*;
 import frc.robot.vision.*;
