+++ Warbots2020/src/main/java/frc/robot/intake/Eject.java
@@ -8,6 +8,7 @@
 package frc.robot.intake;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.subsystems.Intake;
 
 public class Eject extends CommandBase
 {
