+++ Warbots2020/src/main/java/frc/robot/robot/RobotContainer.java
@@ -19,6 +19,11 @@ import frc.robot.climber.*;
 import frc.robot.loader.*;
 import frc.robot.intake.*;
 import frc.robot.shooter.*;
+import frc.robot.subsystems.Climber;
+import frc.robot.subsystems.DriveTrain;
+import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.Loader;
+import frc.robot.subsystems.Shooter;
 import frc.robot.dashboard.*;
 import frc.robot.vision.*;
 import frc.robot.drivetrain.*;
