+++ Warbots2020/src/main/java/frc/robot/shooter/SpinUp.java
@@ -9,6 +9,8 @@ package frc.robot.shooter;
 
 import java.time.LocalDateTime;
 
+import frc.robot.subsystems.Shooter;
+
 public class SpinUp extends ShooterCommand
 {
   public LocalDateTime endTime;
